Renamed Helper.d.ts to ELMTypes.d.ts. Added Statement, Inclue and ValueSet interfaces.
